Added a drawer component.
Added the Material UI Hidden component to hide the stepper and Gathered Information List for medium, small, and extra small screens. Added a drawer component in the component folder, which is used in the simulation window. The stepper components works pefectly inside the drawer, however, the gathered information list needs some work on making the pop ups stay. Also need to work on the formatting of the drawer.